Extract `_blocking` function flavors to a dedicated file
